moved ObjectDescription from xr to contents
